pushing the change to the input system.
The new input system use an AbstractKey binding approach which make the entire InputSystem unified. I am open for another way to approach to make it less obnoxious / annoying to deal with. I feel having the
Input.Shared.mouse.isReleased() could be an approach but it also feel non unified.